corrected task 6

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,5 @@
 # include <stdio.h>
+
 /**
  * main - prints strings in the prinf functions
  *
@@ -14,10 +15,10 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("Size of a char: %ld byte(s)\n", sizeof(c));
-	printf("Size of an int: %ld byte(s)\n", sizeof(i));
-	printf("Size of a long int: %ld byte(s)\n", sizeof(li));
-	printf("size of a long long int: %ld byte(s)\n", sizeof(lli));
- 	printf("Size of a float: %ld byte(s)\n", sizeof(f));
-	return(0);
+	printf("Size of a char: %ld byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %ld byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %ld byte(s)\n", (unsigned long)sizeof(li));
+	printf("size of a long long int: %ld byte(s)\n", (unsigned long)sizeof(lli));
+ 	printf("Size of a float: %ld byte(s)\n", (unsigned long)sizeof(f));
+	return (0);
 }
